nita_cloud_function/requirements.txt
functions-framework==3.3.0
firebase-admin==6.2.0
google-cloud-storage==2.10.0

requests==2.31.0
Flask==2.2.5
lxml==4.9.3
beautifulsoup4==4.12.2